3.93.122
Fixed error on news article update.
Fixed error with comments function and the read status on comments as well.
The calendar on the project page now retains its current view and date when project tasks are updated. This ensures that if you're viewing tasks in a future or past date, the calendar won't reset to the current date after an update, providing a smoother user experience.
Changed itsm modules rights management to Group based instead of Role based.
Improved LDAP user import and improved view on document archive items.